bot/requirements.txt
pyTelegramBotAPI==4.12.0
schedule==1.2.0
python-dotenv==1.0.0
